+++ app/Http/Controllers/JenisEventController.php
@@ -2,10 +2,11 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Event;
+use App\Models\Jenis_Event;
+
 use Illuminate\Http\Request;
 use RealRashid\SweetAlert\Facades\Alert;
-
-use App\Models\Jenis_Event;
 
 class JenisEventController extends Controller
 {
@@ -43,6 +44,13 @@ class JenisEventController extends Controller
     }
 
     public function destroy($id) {
+        $checkChildID = Event::where('jenis_event_id', $id)->count();
+
+        if ($checkChildID > 0) {
+            Alert::error('Gagal Menghapus!', 'Tidak dapat menghapus karena data masih digunakan.');
+            return redirect()->back();
+        }
+
         Jenis_Event::destroy($id);
         toast('Jenis Event Terhapus', 'success');
         return redirect()->back();
